// src/api/generated/models/task.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Shadcn Admin API
 * API for the Shadcn Admin dashboard application
 * OpenAPI spec version: 1.0.0
 */
import type { TaskStatus } from './taskStatus';
import type { TaskLabel } from './taskLabel';
import type { TaskPriority } from './taskPriority';

export interface Task {
  /** Task ID in format TASK-XXXX */
  id: string;
  title: string;
  status: TaskStatus;
  label: TaskLabel;
  priority: TaskPriority;
  createdAt?: string;
  updatedAt?: string;
  assignee?: string;
  description?: string;
  dueDate?: string;
}