// src/api/generated/models/user.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Shadcn Admin API
 * API for the Shadcn Admin dashboard application
 * OpenAPI spec version: 1.0.0
 */
import type { UserStatus } from './userStatus';
import type { UserRole } from './userRole';

export interface User {
  id: string;
  firstName: string;
  lastName: string;
  username: string;
  email: string;
  phoneNumber?: string;
  status: UserStatus;
  role: UserRole;
  createdAt?: string;
  updatedAt?: string;
}