// src/api/generated/models/app.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Shadcn Admin API
 * API for the Shadcn Admin dashboard application
 * OpenAPI spec version: 1.0.0
 */

export interface App {
  id: string;
  name: string;
  desc: string;
  /** SVG or image URL for the app logo */
  logo?: string;
  connected: boolean;
}