// src/api/generated/models/createUserRequest.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Shadcn Admin API
 * API for the Shadcn Admin dashboard application
 * OpenAPI spec version: 1.0.0
 */
import type { UserRole } from './userRole';

export interface CreateUserRequest {
  firstName: string;
  lastName: string;
  email: string;
  phoneNumber?: string;
  role: UserRole;
}