// src/api/generated/models/updateUserRequest.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Shadcn Admin API
 * API for the Shadcn Admin dashboard application
 * OpenAPI spec version: 1.0.0
 */
import type { UserStatus } from './userStatus';
import type { UserRole } from './userRole';

export interface UpdateUserRequest {
  firstName?: string;
  lastName?: string;
  email?: string;
  phoneNumber?: string;
  status?: UserStatus;
  role?: UserRole;
}